fix: move types folder to src
